pyproject.toml
[tool.poetry]
name = "multi-index-demo"
version = "0.1.2"
description = "This repository can be used as a demo for multi-index question answering"
authors = ["christoph_hiemenz <chiemenz@positivethinking.tech>"]
readme = "README.md"
packages = [{include = "multi_index_demo"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
streamlit = "^1.26.0"
llama-index = "^0.8.11.post3"
sentence-transformers = "^2.2.2"
networkx = "^3.1"
scikit-learn = "^1.3.0"
scipy = "^1.11.2"
pandas = "^2.0.3"
numpy = "^1.25.2"
matplotlib = "^3.7.2"
pymupdf = "^1.23.1"
faiss-cpu = "^1.7.4"
seaborn = "^0.12.2"
python-dotenv = "^1.0.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
